refactoring package name processingUsingStrategy to strategy #9
