+++ src/main/java/sudoku/command/ManualInvoker.java
@@ -1,8 +1,8 @@
 package sudoku.command;
 
 import sudoku.model.Sudoku;
-import sudoku.processingUsingStrategy.NakedSingleInACell;
-import sudoku.processingUsingStrategy.Resolvable;
+import sudoku.strategy.NakedSingleInACell;
+import sudoku.strategy.Resolvable;
 import sudoku.stepHandlers.OneChangeStep;
 import sudoku.stepHandlers.Step;
