+++ src/main/java/sudoku/strategy/PointingPairsInCell.java
@@ -1,4 +1,4 @@
-package sudoku.processingUsingStrategy;
+package sudoku.strategy;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
