+++ src/main/java/sudoku/strategy/Solver.java
@@ -1,4 +1,4 @@
-package sudoku.processingUsingStrategy;
+package sudoku.strategy;
 
 import sudoku.model.Sudoku;
