+++ src/test/java/sudoku/strategy/BacktrackLuciaTest.java
@@ -1,4 +1,4 @@
-package sudoku.processingUsingStrategy;
+package sudoku.strategy;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -10,10 +10,11 @@ import sudoku.readers.FileSudokuReader;
 import java.io.File;
 import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
-class HiddenSingleInACellTest {
-    static ClassLoader classLoader = new HiddenSingleInACellTest().getClass().getClassLoader();
+public class BacktrackLuciaTest {
+
+    static ClassLoader classLoader = new BacktrackLuciaTest().getClass().getClassLoader();
 
     private static final String inp1 = new File(classLoader.getResource("inputs/NakedSingleInACell/extremelySimple.txt").getFile()).getPath();
     private static final String out1 = new File(classLoader.getResource("outputs/NakedSingleInACell/extremelySimple.txt").getFile()).getPath();
@@ -22,7 +23,19 @@ class HiddenSingleInACellTest {
     private static final String out2 = new File(classLoader.getResource("outputs/simple1.txt").getFile()).getPath();
 
     private static final String inp3 = new File(classLoader.getResource("inputs/harder1.txt").getFile()).getPath();
-    private static final String out3 = new File(classLoader.getResource("outputs/HiddenSingleInACell/harder1.txt").getFile()).getPath();
+    private static final String out3 = new File(classLoader.getResource("outputs/harder1.txt").getFile()).getPath();
+
+    private static final String inp4 = new File(classLoader.getResource("inputs/extremelyHard1.txt").getFile()).getPath();
+    private static final String out4 = new File(classLoader.getResource("outputs/extremelyHard1.txt").getFile()).getPath();
+
+    private static final String inp5 = new File(classLoader.getResource("inputs/extremelyHard2.txt").getFile()).getPath();
+    private static final String out5 = new File(classLoader.getResource("outputs/extremelyHard2.txt").getFile()).getPath();
+
+    private static final String inp6 = new File(classLoader.getResource("inputs/extremelyHard3.txt").getFile()).getPath();
+    private static final String out6 = new File(classLoader.getResource("outputs/extremelyHard3.txt").getFile()).getPath();
+
+    private static final String inp7 = new File(classLoader.getResource("inputs/extremelyHard3.txt").getFile()).getPath();
+    private static final String out7 = new File(classLoader.getResource("outputs/extremelyHard3.txt").getFile()).getPath();
 
     @ParameterizedTest
     @MethodSource("linksToInputs")
@@ -33,20 +46,8 @@ class HiddenSingleInACellTest {
 
         try {
             Sudoku sudoku = new Sudoku(inputData);
-            NakedSingleInACell nakedSingleInACell = new NakedSingleInACell();
-            HiddenSingleInACell hiddenSingleInACell = new HiddenSingleInACell();
-
-            do {
-                nakedSingleInACell.resolveSudoku(sudoku);
-                System.out.println(" N ");
-                if (!nakedSingleInACell.isUpdated()) {
-                    hiddenSingleInACell.resolveSudoku(sudoku);
-                    System.out.println(" H ");
-                }
-            } while (nakedSingleInACell.isUpdated() || hiddenSingleInACell.isUpdated());
-
-
-
+            BacktrackLucia backtrackLucia = new BacktrackLucia();
+            sudoku = backtrackLucia.resolveSudoku(sudoku);
             printPoss(sudoku);
             System.out.println("=================================");
             assertArrayEquals(expectedOutput, setArrayAccordingToObjectValues(sudoku));
@@ -56,9 +57,13 @@ class HiddenSingleInACellTest {
     }
 
     private static Stream<Arguments> linksToInputs() {
-        return Stream.of(Arguments.of(HiddenSingleInACellTest.inp1, HiddenSingleInACellTest.out1),
-                Arguments.of(HiddenSingleInACellTest.inp2, HiddenSingleInACellTest.out2),
-                Arguments.of(HiddenSingleInACellTest.inp3, HiddenSingleInACellTest.out3)
+        return Stream.of(Arguments.of(BacktrackLuciaTest.inp1, BacktrackLuciaTest.out1),
+                Arguments.of(BacktrackLuciaTest.inp2, BacktrackLuciaTest.out2),
+                Arguments.of(BacktrackLuciaTest.inp3, BacktrackLuciaTest.out3),
+                Arguments.of(BacktrackLuciaTest.inp4, BacktrackLuciaTest.out4),
+                Arguments.of(BacktrackLuciaTest.inp5, BacktrackLuciaTest.out5),
+                Arguments.of(BacktrackLuciaTest.inp6, BacktrackLuciaTest.out6),
+                Arguments.of(BacktrackLuciaTest.inp7, BacktrackLuciaTest.out7)
         );
     }
