+++ src/test/java/sudoku/strategy/NakedSingleInACellTest.java
@@ -1,4 +1,4 @@
-package sudoku.processingUsingStrategy;
+package sudoku.strategy;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -12,31 +12,24 @@ import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 
-public class BacktrackLuciaTest {
+class NakedSingleInACellTest {
 
-    static ClassLoader classLoader = new BacktrackLuciaTest().getClass().getClassLoader();
+    static ClassLoader classLoader = new NakedSingleInACellTest().getClass().getClassLoader();
 
     private static final String inp1 = new File(classLoader.getResource("inputs/NakedSingleInACell/extremelySimple.txt").getFile()).getPath();
     private static final String out1 = new File(classLoader.getResource("outputs/NakedSingleInACell/extremelySimple.txt").getFile()).getPath();
 
     private static final String inp2 = new File(classLoader.getResource("inputs/simple1.txt").getFile()).getPath();
-    private static final String out2 = new File(classLoader.getResource("outputs/simple1.txt").getFile()).getPath();
+    private static final String out2 = new File(classLoader.getResource("outputs/NakedSingleInACell/simple1.txt").getFile()).getPath();
 
     private static final String inp3 = new File(classLoader.getResource("inputs/harder1.txt").getFile()).getPath();
-    private static final String out3 = new File(classLoader.getResource("outputs/harder1.txt").getFile()).getPath();
+    private static final String out3 = new File(classLoader.getResource("outputs/NakedSingleInACell/harder1.txt").getFile()).getPath();
 
-    private static final String inp4 = new File(classLoader.getResource("inputs/extremelyHard1.txt").getFile()).getPath();
-    private static final String out4 = new File(classLoader.getResource("outputs/extremelyHard1.txt").getFile()).getPath();
 
-    private static final String inp5 = new File(classLoader.getResource("inputs/extremelyHard2.txt").getFile()).getPath();
-    private static final String out5 = new File(classLoader.getResource("outputs/extremelyHard2.txt").getFile()).getPath();
-
-    private static final String inp6 = new File(classLoader.getResource("inputs/extremelyHard3.txt").getFile()).getPath();
-    private static final String out6 = new File(classLoader.getResource("outputs/extremelyHard3.txt").getFile()).getPath();
-
-    private static final String inp7 = new File(classLoader.getResource("inputs/extremelyHard3.txt").getFile()).getPath();
-    private static final String out7 = new File(classLoader.getResource("outputs/extremelyHard3.txt").getFile()).getPath();
-
+    /**
+     * extremelySimple.txt input should be completely resolved by NakedSingleInACell() method
+     * hard inputs won't be changed by this method at first at all
+     */
     @ParameterizedTest
     @MethodSource("linksToInputs")
     void resolveSudoku(String inputSudokuMatrixPath, String expectedSudokuOutputPath) {
@@ -46,8 +39,8 @@ public class BacktrackLuciaTest {
 
         try {
             Sudoku sudoku = new Sudoku(inputData);
-            BacktrackLucia backtrackLucia = new BacktrackLucia();
-            sudoku = backtrackLucia.resolveSudoku(sudoku);
+            NakedSingleInACell nakedSingleInACell = new NakedSingleInACell();
+            nakedSingleInACell.resolveSudoku(sudoku);
             printPoss(sudoku);
             System.out.println("=================================");
             assertArrayEquals(expectedOutput, setArrayAccordingToObjectValues(sudoku));
@@ -57,13 +50,9 @@ public class BacktrackLuciaTest {
     }
 
     private static Stream<Arguments> linksToInputs() {
-        return Stream.of(Arguments.of(BacktrackLuciaTest.inp1, BacktrackLuciaTest.out1),
-                Arguments.of(BacktrackLuciaTest.inp2, BacktrackLuciaTest.out2),
-                Arguments.of(BacktrackLuciaTest.inp3, BacktrackLuciaTest.out3),
-                Arguments.of(BacktrackLuciaTest.inp4, BacktrackLuciaTest.out4),
-                Arguments.of(BacktrackLuciaTest.inp5, BacktrackLuciaTest.out5),
-                Arguments.of(BacktrackLuciaTest.inp6, BacktrackLuciaTest.out6),
-                Arguments.of(BacktrackLuciaTest.inp7, BacktrackLuciaTest.out7)
+        return Stream.of(Arguments.of(NakedSingleInACellTest.inp1, NakedSingleInACellTest.out1),
+                Arguments.of(NakedSingleInACellTest.inp2, NakedSingleInACellTest.out2),
+                Arguments.of(NakedSingleInACellTest.inp3, NakedSingleInACellTest.out3)
         );
     }
